+++ packages/jest-util/src/__tests__/createProcessObject.test.js
@@ -6,6 +6,7 @@
  */
 
 import EventEmitter from 'events';
+import isStream from 'is-stream';
 import createProcessObject from '../createProcessObject';
 
 it('creates a process object that looks like the original one', () => {
@@ -22,9 +23,9 @@ it('creates a process object that looks like the original one', () => {
     expect(fakeProcess[key]).not.toBe(process[key]);
   });
 
-  // Check that process.stdout/stderr are the same.
-  expect(process.stdout).toBe(fakeProcess.stdout);
-  expect(process.stderr).toBe(fakeProcess.stderr);
+  // Check that process.stdout/stderr are streams.
+  expect(isStream(fakeProcess.stdout)).toBe(true);
+  expect(isStream(fakeProcess.stderr)).toBe(true);
 });
 
 it('fakes require("process") so it is equal to "global.process"', () => {