+++ types/Console.js
@@ -7,8 +7,10 @@
  * @flow
  */
 
+export type LogApi = 'console' | 'process.stdout' | 'process.stderr';
 export type LogMessage = string;
 export type LogEntry = {|
+  api: LogApi,
   message: LogMessage,
   origin: string,
   type: LogType,
@@ -27,5 +29,6 @@ export type LogType =
   | 'info'
   | 'log'
   | 'time'
-  | 'warn';
+  | 'warn'
+  | 'write';
 export type ConsoleBuffer = Array<LogEntry>;