+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - `[jest-config]` Add `dependencyExtractor` option to use a custom module to extract dependencies from files ([#7313](https://github.com/facebook/jest/pull/7313))
 - `[jest-haste-map]` [**BREAKING**] Expose relative paths when getting the file iterator ([#7321](https://github.com/facebook/jest/pull/7321))
 - `[jest-config]` Add `haste.computeSha1` option to compute the sha-1 of the files in the haste map ([#7345](https://github.com/facebook/jest/pull/7345))
+- `[jest-runner]` [**BREAKING**] Intercept `process.stdout` and `process.stderr` in the same way as `console` methods ([#6524](https://github.com/facebook/jest/pull/6524))
 
 ### Fixes